requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==4.0.5
django-cors-headers==3.13.0
django-debug-toolbar==3.4.0
django-mathfilters==1.0.0
django-pivot==1.9.0
djangorestframework==3.13.1
mysqlclient==2.1.0
numpy==1.23.1
pandas==1.4.3
Pillow==9.1.1
pycodestyle==2.8.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
